refactor: separate `useLazyLoadedContacts`
And add docstring mentioning performance
