+++ src/renderer/components/contact/ContactList.tsx
@@ -167,7 +167,71 @@ export function useContactsNew(listFlags: number, initialQueryStr: string) {
   return [state, search] as [typeof state, typeof search]
 }
 
-export function useContactIds(listFlags: number, queryStr: string | undefined) {
+/**
+ * Why not just `BackendRemote.rpc.getContacts()`?
+ * Because if the user has thousands of contacts (which does happen),
+ * `BackendRemote.rpc.getContacts()` can take a few seconds.
+ * `BackendRemote.rpc.getContactIds()` is much faster
+ * Same for rendering all contacts at the same time,
+ * see {@link ContactList} docsting
+ * See https://github.com/deltachat/deltachat-desktop/issues/1830#issuecomment-2122549915
+ */
+export function useLazyLoadedContacts(
+  listFlags: number,
+  queryStr: string | undefined
+) {
+  const accountId = selectedAccountId()
+  const { contactIds, queryStrIsValidEmail } = useContactIds(
+    listFlags,
+    queryStr
+  )
+
+  const enum LoadStatus {
+    FETCHING = 1,
+    LOADED = 2,
+  }
+
+  // TODO perf: shall we use Map instead of an object?
+  // Or does it not matter since there is not going to be too many contacts?
+  const [contactCache, setContactCache] = useState<{
+    [id: number]: Type.Contact | undefined
+  }>({})
+  const [contactLoadState, setContactLoading] = useState<{
+    [id: number]: undefined | LoadStatus.FETCHING | LoadStatus.LOADED
+  }>({})
+
+  const isContactLoaded: (index: number) => boolean = index =>
+    !!contactLoadState[contactIds[index]]
+  const loadContacts: (
+    startIndex: number,
+    stopIndex: number
+  ) => Promise<void> = async (startIndex, stopIndex) => {
+    const ids = contactIds.slice(startIndex, stopIndex + 1)
+
+    setContactLoading(state => {
+      ids.forEach(id => (state[id] = LoadStatus.FETCHING))
+      return state
+    })
+
+    const contacts = await BackendRemote.rpc.getContactsByIds(accountId, ids)
+    setContactCache(cache => ({ ...cache, ...contacts }))
+    setContactLoading(state => {
+      ids.forEach(id => (state[id] = LoadStatus.LOADED))
+      return state
+    })
+  }
+
+  return {
+    contactIds,
+    isContactLoaded,
+    loadContacts,
+    contactCache,
+
+    queryStrIsValidEmail,
+  }
+}
+
+function useContactIds(listFlags: number, queryStr: string | undefined) {
   const accountId = selectedAccountId()
   const [state, setState] = useState<{
     contactIds: number[]
